Show panic message in fmt::Debug for BlockOnError
Helps with debugging, is still missing the location though, which is a
shame.
